feat(dashboard): add detail and edit athlete pages with dynamic routing and form handling
